requirements.txt
QDarkStyle==3.2.3
pyinstaller==6.9.0
pymongo==4.8.0
dnspython==2.5.0
PyQt6==6.7.1
python-dotenv==1.0.1